Code Changes for: GCP-245 - Implement Get Character Level Values by Observation Unit ID List use case GCP-247 - Implement Get Numeric Level Values by Observation Unit ID List use case
git-svn-id: https://efficio.atlassian.net/svn/GCP/trunk/IBPMiddleware@62 4ebb2e00-0138-4608-bc6e-edf3948d4863

+++ src/main/java/org/generationcp/middleware/pojos/NumericLevelElement.java
@@ -12,10 +12,10 @@ public class NumericLevelElement implements Serializable
 	private Integer ounitId;
 	private Integer factorId;
 	private String factorName;
-	private Integer value;
+	private Double value;
 	
 	public NumericLevelElement(Integer ounitId, Integer factorId,
-			String factorName, Integer value)
+			String factorName, Double value)
 	{
 		super();
 		this.ounitId = ounitId;
@@ -54,12 +54,12 @@ public class NumericLevelElement implements Serializable
 		this.factorName = factorName;
 	}
 
-	public Integer getValue()
+	public Double getValue()
 	{
 		return value;
 	}
 
-	public void setValue(Integer value)
+	public void setValue(Double value)
 	{
 		this.value = value;
 	}
